add CacheFileResolver
