+++ lib/src/models/config/global_cache_config.dart
@@ -5,6 +5,8 @@ import 'package:http_cache_stream/http_cache_stream.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
 
+import 'cache_file_resolver.dart';
+
 /// A configuration class for [HttpCacheManager].
 ///
 /// This class is used to configure the behavior for all [HttpCacheStream] instances,
@@ -26,6 +28,7 @@ class GlobalCacheConfig implements CacheConfiguration {
     this.onCacheDone,
     this.requestTimeout = const Duration(seconds: 60),
     this.readTimeout = const Duration(seconds: 30),
+    this.cacheFileResolver = defaultCacheFileResolver,
   })  : httpClient = customHttpClient ?? Client(),
         requestHeaders = requestHeaders ?? {},
         responseHeaders = responseHeaders ?? {},
@@ -103,6 +106,11 @@ class GlobalCacheConfig implements CacheConfiguration {
 
   @override
   bool saveAllHeaders;
+
+  /// A function that takes the cache directory and source URL, and returns a [File] where the cache should be stored.
+  /// This allows for custom file naming and organization strategies. By default, it generates a file path based on the URL structure.
+  /// This function is called for every cache stream, unless if a custom file is provided when creating the cache stream.
+  final CacheFileResolver cacheFileResolver;
 
   /// Callback function fired when a cache stream download is completed.
   void Function(HttpCacheStream cacheStream, File cacheFile)? onCacheDone;
